refactor(proxy): extract ProxyController creation logic
