+++ sln/src/DotNetTestNSpec/DesignTime/DiscoveryRunner.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using DotNetTestNSpec.Proxy;
+using System;
 
 namespace DotNetTestNSpec.DesignTime
 {
     public class DiscoveryRunner : ITestRunner
     {
-        public DiscoveryRunner(CommandLineOptions commandLineOptions)
+        public DiscoveryRunner(CommandLineOptions commandLineOptions, ProxyFactory proxyFactory)
         {
             // TODO
         }
