+++ sln/src/DotNetTestNSpec/DesignTime/ExecutionRunner.cs
@@ -1,10 +1,11 @@
-﻿using System;
+﻿using DotNetTestNSpec.Proxy;
+using System;
 
 namespace DotNetTestNSpec.DesignTime
 {
     public class ExecutionRunner : ITestRunner
     {
-        public ExecutionRunner(CommandLineOptions commandLineOptions)
+        public ExecutionRunner(CommandLineOptions commandLineOptions, ProxyFactory proxyFactory)
         {
             // TODO
         }
