+++ sln/src/DotNetTestNSpec/RunnerFactory.cs
@@ -1,6 +1,7 @@
 ﻿using DotNetTestNSpec.ConsoleTime;
 using DotNetTestNSpec.DesignTime;
 using DotNetTestNSpec.Parsing;
+using DotNetTestNSpec.Proxy;
 using System;
 
 namespace DotNetTestNSpec
@@ -16,19 +17,21 @@ namespace DotNetTestNSpec
         {
             var commandLineOptions = argumentParser.Parse(args);
 
+            var proxyFactory = new ProxyFactory();
+
             ITestRunner runner;
 
             if (!commandLineOptions.DotNet.DesignTime)
             {
-                runner = new ConsoleRunner(commandLineOptions);
+                runner = new ConsoleRunner(commandLineOptions, proxyFactory);
             }
             else if (commandLineOptions.DotNet.List)
             {
-                runner = new DiscoveryRunner(commandLineOptions);
+                runner = new DiscoveryRunner(commandLineOptions, proxyFactory);
             }
             else if (commandLineOptions.DotNet.WaitCommand)
             {
-                runner = new ExecutionRunner(commandLineOptions);
+                runner = new ExecutionRunner(commandLineOptions, proxyFactory);
             }
             else
             {
